xraygpt_requirements.txt
accelerate==0.15.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.37.0
bleach==6.0.0
blis==0.7.9
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.8
cchardet==2.1.7
certifi @ file:///croot/certifi_1671487769961/work/certif
cffi @ file:///croot/cffi_1670423208954/work
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cmake==3.26.3
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
cymem==2.0.7
dataclasses==0.6
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.0
executing==1.2.0
ExifRead-nocycle==3.0.1
fairscale==0.4.13
fastapi==0.95.1
fastChat==0.1
fastjsonschema==2.16.3
ffmpy==0.3.0
filelock==3.9.0
fire==0.5.0
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fschat @ git+https://github.com/lm-sys/FastChat.git@f34f28cedcb8906fd026f22ec3ef41435a8e24ac
fsspec==2022.11.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
googletrans==3.0.0
gradio==3.23.0
gradio_client==0.0.8
h11==0.9.0
h2==3.2.0
hiq-python==1.1.12
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.13.4
hyperframe==5.2.0
idna==2.10
imageio==2.27.0
img2dataset==1.41.0
importlib-metadata==6.5.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
linkify-it-py==2.0.0
lit==16.0.1
llvmlite==0.39.1
markdown-it-py==2.2.0
markdown2==2.4.8
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
MedCLIP==0.0.3
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numba==0.56.4
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
openai==0.27.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
orjson==3.8.10
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
peft==0.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-itree==0.0.19
pyarrow==12.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
pyllama==0.0.9
pynndescent==0.5.9
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qudida==0.0.4
regex==2022.10.31
requests @ file:///croot/requests_1678709721434/work
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.4
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
semantic-version==2.10.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.19.1
setproctitle==1.3.2
shortuuid==1.0.11
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
starlette==0.26.1
svgwrite==1.4.3
sympy==1.11.1
tenacity==8.2.2
termcolor==2.2.0
terminado==0.17.1
textaugment==1.3.4
textblob==0.17.1
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.2
toolz==0.12.0
torch==2.0.0
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.3
tqdm==4.64.1
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@3d3204c025b6b5de013e07dd364208e28b4d9589
triton==2.0.0
typer==0.7.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
tzdata==2023.3
uc-micro-py==1.0.1
umap-learn==0.5.3
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1680254681959/work
uvicorn==0.21.1
wandb==0.12.21
wasabi==1.1.1
wavedrom==2.0.3.post3
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.2
wget==3.2
xxhash==3.2.0
yarl==1.8.2
zipp==3.14.0
